+++ Example/Tests/WhiteTestSocket.h
@@ -1,0 +1,19 @@
+//
+//  WhiteTestSocket.h
+//  Whiteboard_Tests
+//
+//  Created by xuyunshi on 2022/3/7.
+//  Copyright © 2022 leavesster. All rights reserved.
+//
+
+#import "WhiteSocket.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface WhiteTestSocket : WhiteSocket
+
+@property (nonatomic, strong) NSMutableDictionary *testAbandonMessageDic;
+
+@end
+
+NS_ASSUME_NONNULL_END