+++ Whiteboard/Classes/Room/WhiteRoomConfig.h
@@ -9,29 +9,10 @@
 #import "WhiteMemberInformation.h"
 #import "WhiteCameraBound.h"
 #import "WhiteConsts.h"
-
+#import "WhiteWindowParams.h"
 NS_ASSUME_NONNULL_BEGIN
 
-typedef NSString * WhitePrefersColorScheme NS_STRING_ENUM;
-/** auto只有在iOS13以上才会生效*/
-FOUNDATION_EXPORT WhitePrefersColorScheme const WhitePrefersColorSchemeAuto;
-FOUNDATION_EXPORT WhitePrefersColorScheme const WhitePrefersColorSchemeLight;
-FOUNDATION_EXPORT WhitePrefersColorScheme const WhitePrefersColorSchemeDark;
 
-@interface WhiteWindowParams : WhiteObject
-
-/** 各个端本地显示多窗口内容时，高与宽比例，默认为 9:16。该值应该各个端保持统一，否则会有不可预见的情况。 */
-@property (nonatomic, strong) NSNumber *containerSizeRatio;
-/** 多窗口区域（主窗口）以外的空间显示 PS 棋盘背景，默认 YES */
-@property (nonatomic, assign) BOOL chessboard;
-/** 驼峰形式的 CSS，透传给多窗口时，最小化 div 的 css */
-@property (nonatomic, copy, nullable) NSDictionary *collectorStyles;
-/** 是否在网页控制台打印日志，默认 YES */
-@property (nonatomic, assign) BOOL debug;
-/** 暗黑模式, 本地效果， 不会同步到远端， 默认Light, 设置auto只有在iOS13以上才会生效*/
-@property (nonatomic, copy) WhitePrefersColorScheme prefersColorScheme;
-
-@end
 
 /** 
  配置实时房间的参数。 
@@ -122,6 +103,14 @@ FOUNDATION_EXPORT WhitePrefersColorScheme const WhitePrefersColorSchemeDark;
 @property (nonatomic, assign) BOOL disableEraseImage;
 
 /**
+ 是否关闭 ``insertText`` 与 ``updateText`` 操作权限
+ 
+ - `YES`：不允许``insertText`` 与 ``updateText``。
+ - `NO`：允许``insertText`` 与 ``updateText``。
+ */
+@property (nonatomic, assign) BOOL disableTextOperations;
+
+/**
  本地用户的视角边界。详见 [WhiteCameraBound](WhiteCameraBound)。
  */
 @property (nonatomic, strong, nullable) WhiteCameraBound *cameraBound;
@@ -190,3 +179,7 @@ FOUNDATION_EXPORT WhitePrefersColorScheme const WhitePrefersColorSchemeDark;
 @end
 
 NS_ASSUME_NONNULL_END
+
+#if __has_include("WhiteRoomConfig+FPA.h")
+#import "WhiteRoomConfig+FPA.h"
+#endif