+++ Whiteboard/Classes/Room/WhiteRoomState.m
@@ -7,17 +7,12 @@
 
 #import "WhiteRoomState.h"
 
-WhiteWindowBoxState const WhiteWindowBoxStateNormal = @"normal";
-WhiteWindowBoxState const WhiteWindowBoxStateMini = @"minimized";
-WhiteWindowBoxState const WhiteWindowBoxStateMax = @"maximized";
-
 
 @interface WhiteRoomState ()
 
 @property (nonatomic, strong, nullable, readwrite) WhiteMemberState *memberState;
 @property (nonatomic, strong, nullable, readwrite) WhiteBroadcastState *broadcastState;
 @property (nonatomic, strong, nullable, readwrite) NSNumber *zoomScale;
-@property (nonatomic, copy, nullable, readwrite) WhiteWindowBoxState windowBoxState;
 
 @end